+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Error changing Cost Center after placing order
+
 ## [1.44.10] - 2024-10-07
 
 ### Fixed